fix(motor): model NTMF as DBF_USHORT, not DBF_DOUBLE (R23)
C motorRecord.dbd:789 declares NTMF as DBF_USHORT initial("2"); the
special() clamp compares it as an integer (motorRecord.cc:3093-3100).
motor-rs served it as DbFieldType::Double backed by an f64, so its CA
wire type, value quantization, display precision, and graphic limits
all diverged: a put of 2.5 was stored verbatim (C truncates to 2),
precision reported PREC instead of 0, and graphic limits reported
(1e300,-1e300) instead of (65535,0).

Store ntmf as u16, declare DbFieldType::UShort, read/put as
EpicsValue::UShort. The metadata getters already switch UShort ->
precision 0 / limits (65535,0), so they auto-correct. The two motion
consumers cast u16->f64. Min-2 clamp preserved (R24 will revisit its
Ok-vs-ERROR return separately).

Defect family — anchor DBF_USHORT dbd field vs FieldDesc.dbf_type:
NTMF is the only DBF_USHORT field and the only mismatch (SREV=Long,
RTRY=Short verified correct).

Codex parity audit round 1, Category B.

+++ crates/motor-rs/src/record/command_planner.rs
@@ -1877,7 +1877,7 @@ impl MotorRecord {
         }
 
         let diff = new_dval - self.pos.drbv;
-        let deadband = self.timing.ntmf * (self.retry.bdst.abs() + self.retry.rdbd);
+        let deadband = self.timing.ntmf as f64 * (self.retry.bdst.abs() + self.retry.rdbd);
 
         // C `movn`-block STOP_AXIS gate: opposite direction AND error
         // beyond the NTM deadband AND NTM enabled. The direction sign is
